requirements.txt
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
chained_mode_time_tool==0.4
charset-normalizer==3.4.0
concurrent-log-handler==0.9.23
cryptography~=44.0.1
darkdetect==0.8.0
Deprecated==1.2.14
idna==3.10
logging_tree==1.10
MouseInfo==0.1.3
nb-libs==1.8
nb-log==12.7
nb_filelock==0.9
numpy==2.2.0
opencv-python==4.10.0.84
pillow==11.0.0
portalocker==2.3.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAutoGUI==0.9.54
pycparser==2.22
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pynput==1.7.7
pyperclip==1.9.0
PyQt-Fluent-Widgets==1.8.6
PyQt5==5.15.11
PyQt5-Frameless-Window==0.4.3
PyQt5-Qt5==5.15.2
PyQt5_sip==12.16.0
PyRect==0.2.0
PyScreeze==1.0.1
python-json-logger==0.1.10
pytweening==1.2.0
pytz==2024.2
pywin32==308
requests==2.32.3
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
service-identity==24.2.0
six==1.17.0
tomorrow3==1.1.0
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
wrapt==1.17.0
pyuac~=0.0.3
markdown-it-py~=3.0.0
linkify-it-py~=2.0.3
mdit-py-plugins~=0.4.2
packaging~=24.2
playsound3~=2.5.2
cpufeature~=0.2.1
psutil~=7.0.0